+++ database/seeders/DatabaseSeeder.php
@@ -14,6 +14,11 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call(TecnicoSeeder::class);
-        $this->call(ClienteSeeder::class);
+        $this->call(EncomendaSeeder::class);
+        $this->call(EquipamentoSeeder::class);
+        $this->call(MarcaModeloSeeder::class);
+        $this->call(RMASeeder::class);
+        $this->call(ServicoSeeder::class);
+        $this->call(TecnicoSeeder::class);
     }
 }
